model updated to 2 decimal places

+++ watchlist_app/models.py
@@ -18,7 +18,7 @@ class WatchList(models.Model):
     active = models.BooleanField(default=True)
     created = models.DateTimeField(auto_now_add=True)
     platform = models.ForeignKey(StreamPlatform, on_delete=models.CASCADE, related_name='watchlist', null=True)
-    avg_rating = models.FloatField(default=0)
+    avg_rating = models.DecimalField(max_digits=5, decimal_places=2, default=0.00)
     number_rating = models.IntegerField(default=0)
     
     def __str__(self):
